--- main.py
@@ -1,7 +1,0 @@
-import tkinter as tk
-
-import layout
-
-window = tk.Tk()
-minesweeper = layout.FirstScreen(window)
-window.mainloop()
