+++ src/main.py
@@ -1,0 +1,7 @@
+import tkinter as tk
+
+from screens.first_screen import FirstScreen
+
+window = tk.Tk()
+minesweeper = FirstScreen(window)
+window.mainloop()
